LPS-83808 Move soy renderers to their own module
