+++ modules/apps/frontend-taglib/frontend-taglib-soy/bnd.bnd
@@ -1,9 +1,7 @@
 Bundle-Name: Liferay Frontend Taglib Soy
 Bundle-SymbolicName: com.liferay.frontend.taglib.soy
 Bundle-Version: 3.0.0
-Export-Package:\
-	com.liferay.frontend.taglib.soy.servlet.taglib,\
-	com.liferay.frontend.taglib.soy.servlet.taglib.util
+Export-Package: com.liferay.frontend.taglib.soy.servlet.taglib
 Provide-Capability:\
 	osgi.extender;\
 		osgi.extender="jsp.taglib";\
