+++ modules/apps/frontend-taglib/frontend-taglib-soy/src/main/java/com/liferay/frontend/taglib/soy/servlet/taglib/TemplateRendererTag.java
@@ -16,14 +16,14 @@ package com.liferay.frontend.taglib.soy.servlet.taglib;
 
 import com.liferay.frontend.taglib.soy.internal.util.SoyComponentRendererProvider;
 import com.liferay.frontend.taglib.soy.internal.util.SoyContextFactoryUtil;
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.ComponentDescriptor;
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.SoyComponentRenderer;
 import com.liferay.portal.kernel.theme.PortletDisplay;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.ServerDetector;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.template.soy.renderer.ComponentDescriptor;
+import com.liferay.portal.template.soy.renderer.SoyComponentRenderer;
 import com.liferay.portal.template.soy.utils.SoyContext;
 import com.liferay.taglib.util.ParamAndPropertyAncestorTagImpl;
