--- modules/apps/frontend-taglib/frontend-taglib-soy/src/main/resources/com/liferay/frontend/taglib/soy/servlet/taglib/util/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
