+++ modules/apps/portal-template/portal-template-soy-renderer/src/main/java/com/liferay/portal/template/soy/renderer/internal/SoyComponentRendererImpl.java
@@ -12,11 +12,8 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.soy.internal.util;
+package com.liferay.portal.template.soy.renderer.internal;
 
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.ComponentDescriptor;
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.SoyComponentRenderer;
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.SoyRenderer;
 import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.util.Portal;
 
@@ -28,6 +25,9 @@ import java.util.Map;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.liferay.portal.template.soy.renderer.ComponentDescriptor;
+import com.liferay.portal.template.soy.renderer.SoyComponentRenderer;
+import com.liferay.portal.template.soy.renderer.SoyRenderer;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -38,8 +38,8 @@ import org.osgi.service.component.annotations.Reference;
 public class SoyComponentRendererImpl implements SoyComponentRenderer {
 
 	public void renderSoyComponent(
-			HttpServletRequest request, HttpServletResponse response,
-			ComponentDescriptor componentDescriptor, Map<String, ?> context)
+		HttpServletRequest request, HttpServletResponse response,
+		ComponentDescriptor componentDescriptor, Map<String, ?> context)
 		throws IOException, TemplateException {
 
 		renderSoyComponent(
