+++ modules/apps/portal-template/portal-template-soy-renderer/src/main/java/com/liferay/portal/template/soy/renderer/internal/SoyRendererImpl.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.frontend.taglib.soy.internal.util;
+package com.liferay.portal.template.soy.renderer.internal;
 
-import com.liferay.frontend.taglib.soy.servlet.taglib.util.SoyRenderer;
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateConstants;
 import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.template.TemplateManager;
+import com.liferay.portal.template.soy.renderer.SoyRenderer;
 import com.liferay.portal.template.soy.utils.SoyTemplateResourcesProvider;
 
 import java.io.IOException;
