+++ modules/apps/portal-template/portal-template-soy-renderer/src/main/resources/com/liferay/portal/template/soy/renderer/internal/dependencies/bootstrap.js.tpl
